backend/requirements.txt
fastapi==0.95.1 
uvicorn==0.22.0

SQLAlchemy==2.0.13
python-dotenv==1.0.0

psycopg2-binary==2.9.9

